+++ packages/core/documentation/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.40](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fdocumentation%401.7.39...%40kong-ui-public%2Fdocumentation%401.7.40) (2026-09-24)
+
+**Note:** Version bump only for package @kong-ui-public/documentation
+
+
+
+
+
 ## [1.7.39](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fdocumentation%401.7.38...%40kong-ui-public%2Fdocumentation%401.7.39) (2026-09-22)
 
 **Note:** Version bump only for package @kong-ui-public/documentation